feat: replace zap with logrus
Removed zap for logging and added logrus as the replacement. This
is required in the future as the firecracker Go SDK requires logrus.

Added functions to configure logging and propogate a logger via
a context.

Also added a function that will create the required flags. This will be
used later by the cobra commands.

Signed-off-by: Richard Case <richard@weave.works>