feat: added skeleton firecracker provider
Refactored the existing microvm provider package to remove
the provider registry and lazy init. Initially we will only
have 1 provider and perhaps add an additional one (cloud
hypervisor) in the future and so we don't need a registry.

The functions of the firecracker provider are just skeletons and
return errors in most instances. The functions will be added
over time.

Signed-off-by: Richard Case <richard@weave.works>