feat: Mark VM as deleted
Mark the VM in the spec as deleted and let the reconciliation loop
handle it.

For mocking purpose, I introduced a Clock in ports.Collection.

Changed the event to Update event from Delete event. For the
reconciliation loop, they are the same. Let me know if we still want to
use the Delete event.

fixes #74