fix: dhcp not working for macvtap
When using a macvtap interface the guest vm is making the dhcp request
and an offer of an address is being returned. However, cloud-init fails
on the network configuration and times out.

For macvtap interfaces we should be using the mac address of the host
interface in the Firecracker network interface API request and also in
the cloud-init network configuration. This change does this.

Signed-off-by: Richard Case <richard@weave.works>