Document how to set up network with wireless adapter
This is working with wired or wireless adapters. For wired, it's not
necessary because macvtap can do the trick, but it has issues with
wireless adapters.

The whole quick-start document tries to set up a well separated
environment with reignite specific containerd configuration, thinpool
location. Based on this logic the networking setup follows the same
principles.

This is an issue only with wireless adapters, so this is completely
linux specific as VirtualBox machines have as "wired" connection.
Because of this simple reason, the documentation utilize virsh to create
a network with all the required bridge configuration and iptables rules.

related to #128