Delete MicroVM
Reconciliation loop makes the decision if we want to use the Create,
Update, or Delete plan.

Create plan is used only if State is pending. If the state ever moved
from pending to created, we consider changes as an update, because it's
created even if only partially, the update plan should remove or add
resources.

Delete plan is used only if DeletedAt is not zero.

If it was not a Create or a Delete plan, it's an update plan. In an
ideal case, update should return an empty step list.

Right now, Volume delete is not implemented as a step. We have to check
if it's safe or not, because more than one machine can use the same
image. They are not using the same snapshot, but the snapshot is based
on the image.

Return the number of executed steps from `Actuator.Execute`, that way we
can inspect how many steps we wanted to execute and how many steps we
executed. For now it used only to determine if we have to save or not
in `reconcile`. If we executed zero steps, we don't want to save and
create a new version easy time, because it would be an infinite update
loop.

== Temporary side effect (fix coming soon)

Right now after termimating and deleting the vm, the process stays in
memory for a while as `<defunct>`

  root      916724 12.8  0.0      0     0 pts/4    Z+   17:39   0:03 [firecracker] <defunct>

== Dicsovered bug

Something is off with the check for existing vms. Does not matter what
is the provided value for name, if the ns is the same, the system will
tell you it's already there:

  2 UNKNOWN: creating microvm: microvm spec ns1/mvm10000 already exists

Fixes #65