Fix repo.Get
We are using `store.Walk(fn, filter1, filter2)`, but `Walk` calls
`ParseAll(filters...)` and it returns an `Any` filter. Even the
documentation for the functions says:

    // ParseAll parses each filter in ss and returns a filter that will return true
    // if any filter matches the expression.
    //
    // If no filters are provided, the filter will match anything.
    func ParseAll(ss ...string) (Filter, error) {

To get an `All` filter, we can join the filter strings with `,` because
`parser.selectors` creates an `All` resource for all filters separated
by a `,`.

Fixes #145