feat: add mode spec validation
Model spec validation is provided by github.com/go-playground/validator. Validation tags are added to the model in the core/models/microvm.go file. See the validator docs for all available validation options.

The validator is added to the server as a singleton, which is initialised when NewServer is called. The validator is then called after the model spec after the request model has been converted into the internal microvm model.