Detach firecracker process
This rabbit hole was deep, very deep.

Where to start. I tried to fix the Zombie process issue. It was clear,
it's resolvable only if the process is detached, because we don't want
to create a huge memory pool with the list of processes we created and
set up a hook to clean up on various events. Some of the events are not
even in our control, like the MicroVM dies, in this case we end up with
a Zombie process and (later) `reconcile()` will start a new one and if
it dies, we have two zombie processes.

The solution is to start the Firecracker process detached from the main
process.

== Step 1

Set the `SysProcAttr` of the subprocess to `Setsid: true`, that should
tell the system, if the parent process is terminated, let the subprocess
live alone.

At this point the subprocess is still monitored and managed by the main
process. We can tell the main process to let the child go with
`cmd.Process.Release()`, but to do that, we have to wait for the
process with `cmd.Process.Wait()`. To be honest, I don't know why this
call is required, maybe it can't release if the process is not ready to
be released yet, like it was not loaded in the memory yet or something,
really I have no idea. Wait can be dangerous and can block the flow,
running it in a go routine, we can ignore this issue. As soon as parent
could release the child, this go routine end.

This should be enough, but NOPE. Actually, it's working if the main
process did not get a SIGINT. If it was killed with SIGKILL or simply
crashed, the subprocess can live alone.

== Step 2

We pass all the way down a context and the subprocess gets the same
context. If the context is cancelled, it will kill the subprocess. So
resolve this issue, we have to create a new independent context for
`VMCommandBuilder` by replacing `Build(ctx)` with
`Build(context.TODO())`. Maybe `TODO` is not the right answer, but that
indicates we may have to revisit this later, and it makes my life easier
when I search with `aq todo` to list all the places we potentially have
to go back.

Now, this should be enough, but still NOPE. Actually, it's working, even
if the main process gets a SIGINT from an external source like `kill -u SIGINT`,
the subprocess is not terminated, but if we press `^C` on the main
process, the subprocess will be terminated too.

== Step 3

The only difference between an external SIGINT signal and `^C`, yes it's
`stdin`. Reading Firecracker go sdk and Go source code, the subprocess
should not get the `stdin` from the main process if it's not specified,
but somehow it does. As soon as the `stdin` on the command is set to
`&bytes.Buffer{}`, it's working. The subprocess will not get the SIGINT
from `^C`. I don't know why, I have no idea how the subprocess could
catch the `^C` interrupt event.

This should be enough, and it is enough.

This patch resolves two issues:

1. Run MicroVMs detached (they don't die with reignited)
2. When a MicroVM is deleted, the process does not stay there as a
   zombie.

Major change: no more RunDetached flag

Fixes #139