Delete MicroVM (#144)
* Delete MicroVM

Reconciliation loop makes the decision if we want to use the Create,
Update, or Delete plan.

Create plan is used only if State is pending. If the state ever moved
from pending to created, we consider changes as an update, because it's
created even if only partially, the update plan should remove or add
resources.

Delete plan is used only if DeletedAt is not zero.

If it was not a Create or a Delete plan, it's an update plan. In an
ideal case, update should return an empty step list.

Right now, Volume delete is not implemented as a step. We have to check
if it's safe or not, because more than one machine can use the same
image. They are not using the same snapshot, but the snapshot is based
on the image.

Return the number of executed steps from `Actuator.Execute`, that way we
can inspect how many steps we wanted to execute and how many steps we
executed. For now it used only to determine if we have to save or not
in `reconcile`. If we executed zero steps, we don't want to save and
create a new version easy time, because it would be an infinite update
loop.

== Temporary side effect (fix coming soon)

Right now after termimating and deleting the vm, the process stays in
memory for a while as `<defunct>`

  root      916724 12.8  0.0      0     0 pts/4    Z+   17:39   0:03 [firecracker] <defunct>

== Dicsovered bug

Something is off with the check for existing vms. Does not matter what
is the provided value for name, if the ns is the same, the system will
tell you it's already there:

  2 UNKNOWN: creating microvm: microvm spec ns1/mvm10000 already exists

Fixes #65

* Fix repo.Get

We are using `store.Walk(fn, filter1, filter2)`, but `Walk` calls
`ParseAll(filters...)` and it returns an `Any` filter. Even the
documentation for the functions says:

    // ParseAll parses each filter in ss and returns a filter that will return true
    // if any filter matches the expression.
    //
    // If no filters are provided, the filter will match anything.
    func ParseAll(ss ...string) (Filter, error) {

To get an `All` filter, we can join the filter strings with `,` because
`parser.selectors` creates an `All` resource for all filters separated
by a `,`.

Fixes #145

* Release Lease

* Register reflection and helper has script

With `hack/scripts/send.sh`, we can send predefined (and working)
payloads to various endpoints with a simple call. I like BloomRPC, but I
like my tmux and neovim more.

Here is an example how I test the delete workflow:

    ./hack/scripts/send.sh -m CreateMicroVM && sleep 5 ./hack/scripts/send.sh -m DeleteMicroVM

* Detach firecracker process

This rabbit hole was deep, very deep.

Where to start. I tried to fix the Zombie process issue. It was clear,
it's resolvable only if the process is detached, because we don't want
to create a huge memory pool with the list of processes we created and
set up a hook to clean up on various events. Some of the events are not
even in our control, like the MicroVM dies, in this case we end up with
a Zombie process and (later) `reconcile()` will start a new one and if
it dies, we have two zombie processes.

The solution is to start the Firecracker process detached from the main
process.

== Step 1

Set the `SysProcAttr` of the subprocess to `Setsid: true`, that should
tell the system, if the parent process is terminated, let the subprocess
live alone.

At this point the subprocess is still monitored and managed by the main
process. We can tell the main process to let the child go with
`cmd.Process.Release()`, but to do that, we have to wait for the
process with `cmd.Process.Wait()`. To be honest, I don't know why this
call is required, maybe it can't release if the process is not ready to
be released yet, like it was not loaded in the memory yet or something,
really I have no idea. Wait can be dangerous and can block the flow,
running it in a go routine, we can ignore this issue. As soon as parent
could release the child, this go routine end.

This should be enough, but NOPE. Actually, it's working if the main
process did not get a SIGINT. If it was killed with SIGKILL or simply
crashed, the subprocess can live alone.

== Step 2

We pass all the way down a context and the subprocess gets the same
context. If the context is cancelled, it will kill the subprocess. So
resolve this issue, we have to create a new independent context for
`VMCommandBuilder` by replacing `Build(ctx)` with
`Build(context.TODO())`. Maybe `TODO` is not the right answer, but that
indicates we may have to revisit this later, and it makes my life easier
when I search with `aq todo` to list all the places we potentially have
to go back.

Now, this should be enough, but still NOPE. Actually, it's working, even
if the main process gets a SIGINT from an external source like `kill -u SIGINT`,
the subprocess is not terminated, but if we press `^C` on the main
process, the subprocess will be terminated too.

== Step 3

The only difference between an external SIGINT signal and `^C`, yes it's
`stdin`. Reading Firecracker go sdk and Go source code, the subprocess
should not get the `stdin` from the main process if it's not specified,
but somehow it does. As soon as the `stdin` on the command is set to
`&bytes.Buffer{}`, it's working. The subprocess will not get the SIGINT
from `^C`. I don't know why, I have no idea how the subprocess could
catch the `^C` interrupt event.

This should be enough, and it is enough.

This patch resolves two issues:

1. Run MicroVMs detached (they don't die with reignited)
2. When a MicroVM is deleted, the process does not stay there as a
   zombie.

Major change: no more RunDetached flag

Fixes #139

* MaximumRetry and feedback

* Address some of the comments
* Make MaximumRetry configurable
* Emmit MicroVMSpecDeleted event
* No reconciliation on Delete event
* Default detached mode = true

* Testing

* Create plan test
* Delete plan test

* core/steps/runtime test

* Remove unused private method

* Typo fix