Introduce a computed status of a microvm (#150)
* refactor: Change IsRunning into a state check

This change removed the `IsRunning` from the MicroVMService interface
and replaces it with a `Sate` function that returns an enum that can
represent a number of different states for the microvm.

The reasoning behind this is that we need to distinguish between
Firecracker is running but the guest isn't with the scenario where
Firecracker is running and the guest is. The reason is that we have
separate steps to configure Firecracker and start the guest OS and these
steps need separate checks in the `ShouldDo`. When trying to use the the
`IsRunning` for both it caused the paln to not start which would then
cause a re-plan and we would try and configure Firecracker again which
would cause an error: `Resource Busy`.

The `State` that is returned is determined in realtime based on the
status of Firecracker and doesn't come from the spec.

Additionally, the final parts of the delete have been removed from
reconcile into the delete plan. This has meant creating 2 new steps.

Signed-off-by: Richard Case <richard@weave.works>

Co-authored-by: Balazs Nadasdi <yitsushi@gmail.com>