Kill Firecracker if SendCtrlAltDel failed
Firecracker on arm machines does not support the SendCtrlAltDel action
and luckily it throws back an error instead of silently does nothing.

Why lucky? That way if the SendCtrlAltDel failed, we can rety to kill it
with a graceful SIGINT signal and if it did not work, we can throw back
an error and say "we did what we could".

Error message from Firecracker:

    {"fault_message":"SendCtrlAltDel does not supported on aarch64."}

Fixes #152