Disable root volume validation (#220)
The firecracker docs say that only you can only specific either a root
volume or an initrd. However, when we remove the root volume and only
have initrd the mvm will not start.

We are not entirely sure what is happening yet, so for now we are going
to disable the validation and continue using both a root vol and
initrd. It's more of a guideline than an actual rule anyway ;)

There is a separate ticket to track the investigation and potential fix
here: https://github.com/weaveworks/flintlock/issues/219

(Also added a missing test, the removed bit also was not covered so we
should remember to add one when/if we put it back)