Add thinpool setup to general test Setup/Teardown
I had this being done in bash via a Docker Entrypoint + `trap`
combination, but it was quite brittle, and sometimes the teardown would
not run which then meant you had to run a manual step.

This changes brings that into the test so the setup and teardown can be
handled programatically.