Add thinpool setup to general test Setup/Teardown (#223)
I had this being done in bash via a Docker Entrypoint + `trap`
combination, but it was quite brittle, and sometimes the teardown would
not run which then meant you had to run a manual step to remove thinpools
and loop devices which was annoying.

This changes brings that into the test so the setup and teardown can be
handled programatically.

There is still a case where some manual cleanup may be required: when the tests
fail and devices cannot be removed. I have a thing on my list to have an aggressive
cleanup of all mvms etc on failure. Coming soon!

There is also a case where some "child" pool devices are not cleaned up. I have deliberately
not done anything about this as we should actually fix the code that is leaving those lying
around after a failed delete.