fix: set mac address for macvtap if specified
There is a problem with using an auto generated mac address with a macvtap interface. The
kernel appears to change the mac address from what it was originally
created with. The firecracker and cloud init configuration contain the aoriginal mac address.
When the guest boots and cloud-init gets to configured the network the
mac address has changed and this then causes the network configuration to fail as it can't
match on the mac address.

We need to find actual evidence of the above but this has been a known
behaviour with other types of interface and for the those the advice is
to specify a mac address when you create the interface.

This change will use the `guest_mac` if its specified for the mac
address on the host if the interface type is macvtap. For macvtap the
mac address on the host is the same as in the guest.

Signed-off-by: Richard Case <richard@weave.works>