Add config parser class to e2e tool
A new instance of Config will be initialised with a set of default
config. These are just the values I happened to be hardcoding before,
and there was no special reason for any of them, but they work.

Config can then be loaded from a file and validated against a schema.
This config will then overwrite the appropriate values in the object
config.

Config can also be altered via flags. The 'final' config at that point
can be validated.

The validation is all a bit naive right now as I didn't want to get too
into it.