Add thinpool name flag to e2e tests
Adds a new -thinpool custom flag to the e2e tests. This
lets users set a thinpool name, useful for if they want to use
one they made earlier or don't want to conflict with an
existing one.

If using with an existing one, users should set the
-skip.setup.thinpool flag so they don't have some random
pool made