Alter skip dmsetup behaviour in e2e tool
Previously `skip_dmsetup` was under `test` in the tool config file. This
was to disable the loopdevice thinpool setup done by the e2e test suite.
Disabling the test level dmsetup should always be True when running in
Equinix, so I have hardcoded it.

There may be a case in which the runner wants to disable the direct_lvm
thinpool setup, so that option is now available and grouped under
`device` in the config as it is relevant to the create command, not just
running the tests.