feat(test): Wire e2e tests to use direct lvm thinpool for containerd (#270)
* Add thinpool name flag to e2e tests

Adds a new -thinpool custom flag to the e2e tests. This
lets users set a thinpool name, useful for if they want to use
one they made earlier or don't want to conflict with an
existing one.

If using with an existing one, users should set the
-skip.setup.thinpool flag so they don't have some random
pool made

* Alter skip dmsetup behaviour in e2e tool

Previously `skip_dmsetup` was under `test` in the tool config file. This
was to disable the loopdevice thinpool setup done by the e2e test suite.
Disabling the test level dmsetup should always be True when running in
Equinix, so I have hardcoded it.

There may be a case in which the runner wants to disable the direct_lvm
thinpool setup, so that option is now available and grouped under
`device` in the config as it is relevant to the create command, not just
running the tests.