Check capacity of facility before creating device (#272)
This changes the config so that a list of facilities will be tested to
ensure that there is capacity for device creation.

This gives us some resilience in the nightly e2e tests, although if
there is no capacity anywhere then they will still fail.

I had wanted to use a list of metros instead, which would have left
Equinix to find an appropriate facility, but the package I am using to
hit the Equinix API has a bug in that method. I have removed setting the
metro as an option as I did not feel like dealing with that logic right
now. Also I am not really sure which takes precedence if you set both on
the request. Metro I think?