Add opts to configure rep and branch
Just a nice to have when creating the device.

Useful for running the e2es on a branch, otherwise it will just be main
cloned down and tested.

For more flexibility I am going to add a further option which will scp a
built binary (or a tar of the local repo can't decide yet) to make local
dev turnaround even faster.