Add action to bump version of firecracker in e2es
Note: The firecracker version included in the bootstrap.sh of this PR
is a dummy version so I can see that everything works.

The action will open a PR to bump the version when a new one is
available.
The new firecracker release wont actually exist yet, as there are some
manual steps required to update the macvtap feature branch on our fork.
These manual steps will be detailed in the PR.