refactor: cloudinit pointer removal (#316)
A small refactor to the **cloudinit** package to remove pointers where
possible so that its more user friendly to consume.

In the struct definitions we have tried to remove usage of pointers and instead rely on the behavior of json
marshalling and `omitempty`. Its slightly tricky with boolean values so for these we use *bool vs bool. The
reason is that the default value for some cloudinit values is true which isn't the same as the default for bool.

Signed-off-by: Richard Case <richard@weave.works>