feat: bind flintlock server to internal ip
This is for the provisioning script. There will be docs added.

There is an option to not do this, but by default the provision script
will assume you want a VLAN type setup.

There is a hideous little func which is sedding the flintlock service
file, and the service file itself does not look to pretty, but these are
temporary measures just for now.