fix: remove the `imageExists` check before Pull (#359)
Containerd does not need us to check for the existence of an image for
it; it will download any layers which do not exist and wont re-download
the same image twice.

This check is causing us trouble now because with our capmvm base images
we push new shas to the same tags. When we run this `imageExists` check
using the ref (name+label) we don't get new changes on that version.
This fix just leaves containerd to get on with it.