feat: Use flintlock unique id (#327)
* feat: Use flintlock unique id

This commit message will be a QA like ocmmit message

Why in the spec and not somewhere else?

Historically we are using the same spec in create request an responses,
we do not return with status on create request because it's mostly
empty. The uuid could be there, but on create there is no status yet.

Why is it set on create request?

We can't return the uuid in the response if it's not generated and
without the uuid they can't delete the resource, it would be a disaster
if capmvm sends a request to create ns1/mvm1, then it wants to delete it
even before we created the resources. If the delete request would be
only ns1/mvm1 it would delete the old one, or basically any other
microvms that has the same name and namespace (coming from other
management clusters).

Why is it global and not microvm specific?

It seems logical and generating and maintaining a unique id shouldn't be
the job of the microvm provider, their responsibility should be to
manage microvms, metadata and other stuffs are provider independent.

Fixes #291

rename field to be uid instead of uuid because it is an ulid now

remove id and ns from get and delete endpoints

* fix e2e

* Update core/application/commands.go

Co-authored-by: Claudia <claudiaberesford@gmail.com>

* address comments

Co-authored-by: Claudia <claudiaberesford@gmail.com>