fix: failed to create microvm if machine id is too long (#406)
This changes the firecracker provider to use the unique id on the VMID
for the id of the firecracker instance instead of using the
namespace/name combination. The `--id` that you supply to a firecracker
instance can only be a maximum of 64 characters. This is easily exceeded
witgh the namespace and machine name combination. We use a ULID for the
machine unique identifer internally and this is 26 characters long.

Signed-off-by: Richard Case <richard@weave.works>