feat: Add (very) basic auth (#462)
This commit adds some very basic token-based auth to the flintlock
server.

```
flintlockd run --basic-auth-token foo
```

Note that this is not a bearer token, or a username+password combo. It
is simply a token which will be compared against an **encoded** token on
all server calls.

Clients are responsible for adding an `authorization: basic <encoded
token>` header to all calls.

TLS is next on the list.