fix: tap interface attach to bridge (#479)
When requesting that a tap device is added to a microvm then we should
attach (i.e. set the master) to a bridge.

The bridge will be supplied via a new bridge-name flag. And the attach
should not be done for the mmds network interface.

The grpc api has been updated to allow the consumer to specify a
bridge-name that is different as part of the create microvm call.

Signed-off-by: Richard Case <richard.case@outlook.com>