refactor: allow per create specifying provider
Based on feedback to was though we should allow flintlock to use both
microvm providers at the same time. This refactor allows the user to set
a default provider when they start flintlock, and this default provider
will be used if there is no provider name specified in the create
request.

This is still wip and so tests don't pass.

Signed-off-by: Richard Case <richard.case@outlook.com>