refactor: changes required for Firecracker v1.1 Support
When testing with Firecracker v1.1 changes are required for Flintlock to
run. We have encountered issues with MMDS and cloud-init after upgrading
Firecracker from v0.25.2, specifically it looks lik the MMDS isn't
contactable from the guest. An issue has been opened in the Firecracker
repo and until that is resolved we have made changes so that we can
still use Firecracker v1.1.

Specifically we have made these changes:

- A new flag has been introduced to Flintlock that indicates that
  cloud-init should be done from an attached volume and not via the
  MMDS. This has required that we introduce new steps into the
  create/update plan that will create the disk and then mount it into
  the vm.
- As the MMDS cannot be relied upon we have added a feature that will
  allow non-cloudinit metadata to be mounted into the microvm via an
  attached volume.

NOTE: we have made breaking changes to the grpc api with this.

Signed-off-by: Richard Case <richard.case@outlook.com>