feat: Add list all functionality (#489)
This removes the requirement that a call to GetAllMicroVM must have a
namespace argument in the request. We do filtering on the Repo, so it is
not harmful to have all params be optional.

If a call to GetAllMicroVM is made without namespace or name, then all
mvms on the host will be returned.

If a call to GetAllMicroVM is made with just name, then all mvms
matching that name in any namespace will be returned.

If a call to to GetAllMicroVM is made with just namespace, then all mvms
under that namespace will be returned.

We can consider RBAC implications when we get to that.