fix: Pin containerd to 1.6.6
Just temporarily. The 1.6.7 release requires a newer gcc version which
means we would have to bump our ubuntu version (to get the newer gcc
easily) which comes with a few challenges. It may not be much of a big
deal at flintlock's level, but I wanted to get the nightlies working
while I figure it out.

See https://github.com/weaveworks-liquidmetal/liquid-metal-acceptance-tests/pull/11