feat: add pprof to flintlock
This change adds the ability to make pprof available via http. By
default it won't be available. If you specify a debug endpoint via
`--debug-endpoint` then it will start a webserver listening at that
endpoint and pprof is available from there.

We decided to use an endpoint as that will allow users to only expose
the debug endpoint locally or via the network.

Signed-off-by: Richard Case <richard.case@outlook.com>