feat: add support for Firecracker v1+
A small number of changes to support Firecracker v1+. We are sticking
with the v1 MMDS initially andf will look to upgrade to the v2 MMDS in
the future.

To use this version with macvtap we will need to use our fork with the
1.1 changes.

NOTE: incorrect serialization caused the issues with MMDS not working.
Strangely Firecracker didn't complain about the invalid json config.
Time and stress wasted on wht was essentially a type.

Signed-off-by: Richard Case <richard.case@outlook.com>