feat: add new disk service
This change adds a new port for a "disk service" that will be used to
create disk images (i.e. img/iso files). An infrastructure
implemenetation has been added using godisk.

Note: this isn't used by anything in this change. But it will be used in
the future with the Cloud Hypervisor implementation.

Signed-off-by: Richard Case <richard.case@outlook.com>