fix: Set address based on interface in provision.sh
In machines with less simple or "clean" networking, the provision script
could discover multiple `default` interfaces and multiple internal ips.
This led to a messed up config and obviously flintlockd would not start.

This commit changes the script so that if more than one is found, just
the first `default` interface is set.

The found interface is then used to find a single associated internal address.

There are flags to override both these options, which is what most
people will be doing, but the default behaviour needs to work too.