Update provision.sh to start the server with auth (#557)
This also presents the option to pass a full flintlock config file to the script,
allowing users to override defaults and set values without the script having to
expose the exact same flags as flintlock. This avoids a lot of messy expansion
in the script.
Users can thus create certs and put them in a config file, passing that file to
the script: `provision.sh -f flint.cfg`.