refactor: removed `gw` command (#627)
The separate `gw` command has been removed and its functionality
incorporated into the `run` command. The HTTP gateway is then enabled
using `--enable-http`.

NOTE: the code to start the HTTP gateway remains the same. We will need
to refactor it in the future to take into account the the TLS flags.

Signed-off-by: Richard Case <richard.case@outlook.com>
Co-authored-by: Balazs Nadasdi <balazs@weave.works>