test: Update network-interface config for e2es (#664)
This was previously `eth0` which led to a weird and misleading error from
firecracker complaining about `network_interfaces` missing.. when the
api actually accepts `network-interfaces` with a dash. This is a
red-herring: really firecracker just wasn't happy with us setting `eth0`
because it apparently uses that for something else.

I also updated the images used in the test to use our official builds.